requirements.txt
python-dotenv>=1.0.0
azure-identity>=1.14.0
